created Procfile for heroku and added node and yarn version to package json. fixed Bounty and BountyCollection type for reward amount

+++ src/app/types/bounty/Bounty.ts
@@ -1,4 +1,4 @@
-import { Int32, ObjectId } from 'mongodb';
+import { Double, Int32, ObjectId } from 'mongodb';
 
 export interface Bounty {
 	_id?: ObjectId,
@@ -29,7 +29,7 @@ export type UserObject = {
 
 export type Reward = {
 	currency: string,
-	amount: Int32,
+	amount: Double,
 	scale: Int32,
 	amountWithoutScale: Int32,
 };
